+++ aisearch/src/main/java/com/elastic/aisearch/service/WeatherService.java
@@ -29,7 +29,7 @@ public class WeatherService {
 
         return new WeatherDTO(
                 json.path("name").asText(),
-                json.path("main").path("temp").asDouble(),
+                json.path("main").path("temp").asInt(),
                 json.path("weather").get(0).path("description").asText()
         );
     }
